+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/baal/Documents/Web/React/myReactApp/src/index.js":"1","/Users/baal/Documents/Web/React/myReactApp/src/App.js":"2","/Users/baal/Documents/Web/React/myReactApp/src/components/Header/index.js":"3","/Users/baal/Documents/Web/React/myReactApp/src/components/Footer/index.js":"4","/Users/baal/Documents/Web/React/myReactApp/src/components/Layout/index.js":"5","/Users/baal/Documents/Web/React/myReactApp/src/components/PokemonCard/index.js":"6"},{"size":175,"mtime":1611924648635,"results":"7","hashOfConfig":"8"},{"size":4284,"mtime":1611935543045,"results":"9","hashOfConfig":"8"},{"size":293,"mtime":1611912586158,"results":"10","hashOfConfig":"8"},{"size":243,"mtime":1611912586152,"results":"11","hashOfConfig":"8"},{"size":629,"mtime":1611929999642,"results":"12","hashOfConfig":"8"},{"size":1518,"mtime":1611936442316,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"1tqf5u4",{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/baal/Documents/Web/React/myReactApp/src/index.js",[],["27","28"],"/Users/baal/Documents/Web/React/myReactApp/src/App.js",[],"/Users/baal/Documents/Web/React/myReactApp/src/components/Header/index.js",[],"/Users/baal/Documents/Web/React/myReactApp/src/components/Footer/index.js",[],"/Users/baal/Documents/Web/React/myReactApp/src/components/Layout/index.js",[],"/Users/baal/Documents/Web/React/myReactApp/src/components/PokemonCard/index.js",[],{"ruleId":"29","replacedBy":"30"},{"ruleId":"31","replacedBy":"32"},"no-native-reassign",["33"],"no-negated-in-lhs",["34"],"no-global-assign","no-unsafe-negation"]
+[{"/Users/baal/Documents/Web/React/myReactApp/src/index.js":"1","/Users/baal/Documents/Web/React/myReactApp/src/App.js":"2","/Users/baal/Documents/Web/React/myReactApp/src/components/Header/index.js":"3","/Users/baal/Documents/Web/React/myReactApp/src/components/Footer/index.js":"4","/Users/baal/Documents/Web/React/myReactApp/src/components/Layout/index.js":"5","/Users/baal/Documents/Web/React/myReactApp/src/components/PokemonCard/index.js":"6","/Users/baal/Documents/Web/React/myReactApp/src/routes/Game/index.js":"7","/Users/baal/Documents/Web/React/myReactApp/src/routes/Home/index.js":"8","/Users/baal/Documents/Web/React/myReactApp/src/components/MenuHeader/index.js":"9","/Users/baal/Documents/Web/React/myReactApp/src/components/Menu/index.js":"10","/Users/baal/Documents/Web/React/myReactApp/src/components/Navbar/index.js":"11"},{"size":175,"mtime":1612185797554,"results":"12","hashOfConfig":"13"},{"size":506,"mtime":1612228850712,"results":"14","hashOfConfig":"13"},{"size":450,"mtime":1612248400073,"results":"15","hashOfConfig":"13"},{"size":256,"mtime":1612238896728,"results":"16","hashOfConfig":"13"},{"size":648,"mtime":1612183083616,"results":"17","hashOfConfig":"13"},{"size":1528,"mtime":1612189401384,"results":"18","hashOfConfig":"13"},{"size":334,"mtime":1612191879687,"results":"19","hashOfConfig":"13"},{"size":4589,"mtime":1612248399154,"results":"20","hashOfConfig":"13"},{"size":694,"mtime":1612248347417,"results":"21","hashOfConfig":"13"},{"size":700,"mtime":1612246843527,"results":"22","hashOfConfig":"13"},{"size":666,"mtime":1612248362809,"results":"23","hashOfConfig":"13"},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},"1tqf5u4",{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"39","messages":"40","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"41","messages":"42","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"43","messages":"44","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"45","messages":"46","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/Users/baal/Documents/Web/React/myReactApp/src/index.js",[],["47","48"],"/Users/baal/Documents/Web/React/myReactApp/src/App.js",[],"/Users/baal/Documents/Web/React/myReactApp/src/components/Header/index.js",[],"/Users/baal/Documents/Web/React/myReactApp/src/components/Footer/index.js",[],"/Users/baal/Documents/Web/React/myReactApp/src/components/Layout/index.js",[],"/Users/baal/Documents/Web/React/myReactApp/src/components/PokemonCard/index.js",[],"/Users/baal/Documents/Web/React/myReactApp/src/routes/Game/index.js",[],"/Users/baal/Documents/Web/React/myReactApp/src/routes/Home/index.js",[],"/Users/baal/Documents/Web/React/myReactApp/src/components/MenuHeader/index.js",[],"/Users/baal/Documents/Web/React/myReactApp/src/components/Menu/index.js",[],"/Users/baal/Documents/Web/React/myReactApp/src/components/Navbar/index.js",["49"],{"ruleId":"50","replacedBy":"51"},{"ruleId":"52","replacedBy":"53"},{"ruleId":"54","severity":1,"message":"55","line":26,"column":5,"nodeType":"56","endLine":26,"endColumn":106},"no-native-reassign",["57"],"no-negated-in-lhs",["58"],"jsx-a11y/anchor-is-valid","The href attribute is required for an anchor to be keyboard accessible. Provide a valid, navigable address as the href value. If you cannot provide an href, but still need the element to resemble a link, use a button and change it with appropriate styles. Learn more: https://github.com/evcohen/eslint-plugin-jsx-a11y/blob/master/docs/rules/anchor-is-valid.md","JSXOpeningElement","no-global-assign","no-unsafe-negation"]
