+++ src/components/Footer/index.js
@@ -5,7 +5,7 @@ import style from './style.module.css';
 
 const Footer = () => {
 	return (
-		<footer >
+		<footer id="contact" >
 			<div className={style.wrapper}>
 				<h3>THANKS FOR VISITING</h3>
 				<p>© 2021 #ReactMarathon.</p>
