+++ src/components/Header/index.js
@@ -1,14 +1,23 @@
-
 import style from './style.module.css';
 
 
-const Header = ({ title, descr }) => {
+const Header = ({ title, descr, onClickButton }) => {
+	const handleClick = () => {
+
+		onClickButton && onClickButton('game');
+
+	}
+
 	return (
 		<header className={style.root}>
 			<div className={style.forest}></div>
 			<div className={style.container}>
 				<h1>{title}</h1>
 				<p>{descr}</p>
+				<button onClick={handleClick}>
+					Start Game
+				</button>
+
 			</div>
 		</header>
 	)
