+++ src/components/Header/style.module.css
@@ -88,3 +88,28 @@
 	color: #fafafa;
 	line-height: 32px;
 }
+
+.container button {
+	appearance: none;
+	border: 0;
+	border-radius: 5px;
+	background: #9909ac;
+	color: #fff;
+	padding: 8px 16px;
+	font-size: 28px;
+	margin: 20px 0;
+	padding-left: 20px;
+	padding-right: 20px;
+	width: 100%;
+	text-align: center;
+}
+.container button:hover {
+	background: #fff;
+	color: #9909ac;
+}
+
+.container button:focus {
+	outline: none;
+	box-shadow: 0 0 0 4px #640670;
+	/* border: 5px solid #181d23; */
+}
