+++ src/components/Layout/index.js
@@ -1,5 +1,5 @@
 import style from './style.module.css';
-
+import cn from 'classnames'
 
 const Layout = ({ id, title, urlBg, colorBg, children }) => {
 	const styleBg = {}
@@ -13,7 +13,7 @@ const Layout = ({ id, title, urlBg, colorBg, children }) => {
 						<h3>{title}</h3>
 						<span className={style.separator}></span>
 					</div>
-					<div className={[style.desc, style.full].join(" ")}>
+					<div className={cn(style.desc, style.full)}>
 						{children}
 					</div>
 				</article >
