+++ src/components/PokemonCard/index.js
@@ -1,27 +1,30 @@
-import {useState} from 'react';
+import { useState } from 'react';
+
+import cn from 'classnames';
 
 import style from './style.module.css';
 import cardBackSide from './assets/card-back-side.jpg';
 
 
 
-const PokemonCard = ({name, img, id, type, values}) => {
-	
+const PokemonCard = ({ name, img, id, type, values }) => {
+
 	const [isActive, setActive] = useState(false);
 	const handleClick = () => {
 		setActive(!isActive);
 	}
 	return (
 		<div className={style.root} onClick={handleClick}>
-			<div className={`${style.pokemonCard} ${isActive ? style.active : ''}`}>
+
+			<div className={cn(style.pokemonCard, { [style.active]: isActive })}>
 				<div className={style.cardFront}>
-					<div className={`${style.wrap} ${style.front}`}>
-						<div className={`${style.pokemon} ${style[type]}`}>
+					<div className={cn(style.wrap, style.front)}>
+						<div className={cn(style.pokemon, style[type])}>
 							<div className={style.values}>
-								<div className={`${style.count} ${style.top}`}>{values.top}</div>
-								<div className={`${style.count} ${style.right}`}>{values.right}</div>
-								<div className={`${style.count} ${style.bottom}`}>{values.bottom}</div>
-								<div className={`${style.count} ${style.left}`}>{values.left}</div>
+								<div className={cn(style.count, style.top)}>{values.top}</div>
+								<div className={cn(style.count, style.right)}>{values.right}</div>
+								<div className={cn(style.count, style.bottom)}>{values.bottom}</div>
+								<div className={cn(style.count, style.left)}>{values.left}</div>
 							</div>
 							<div className={style.imgContainer}>
 								<img src={img} alt={name} />
@@ -36,7 +39,7 @@ const PokemonCard = ({name, img, id, type, values}) => {
 				</div>
 
 				<div className={style.cardBack}>
-					<div className={`${style.wrap} ${style.back}`}>
+					<div className={cn(style.wrap, style.back)}>
 						<img src={cardBackSide} alt="Сard Backed" />
 					</div>
 				</div>
