+++ src/routes/Game/style.module.css
@@ -1,0 +1,67 @@
+.flex {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	flex-wrap: wrap;
+	align-content: flex-start;
+}
+.container {
+	width: fit-content;
+	position: absolute;
+	right: 0;
+	top: 50%;
+	left: 50%;
+	opacity: 0;
+	transform: translate(-50%, -50%);
+	animation: text-fade-in 1000ms 800ms forwards;
+}
+
+@keyframes text-fade-in {
+	from {
+		right: 0;
+	}
+	to {
+		opacity: 1;
+		right: 25%;
+	}
+}
+
+.container h1 {
+	font-size: 65px;
+	font-weight: normal;
+	font-style: italic;
+	color: #ba13c0;
+	line-height: 65px;
+}
+
+.container p {
+	font-size: 28px;
+	font-weight: normal;
+	color: #fafafa;
+	line-height: 32px;
+}
+
+.container button {
+	appearance: none;
+	border: 0;
+	border-radius: 5px;
+	background: #9909ac;
+	color: #fff;
+	padding: 8px 16px;
+	font-size: 28px;
+	margin: 20px 0;
+	padding-left: 20px;
+	padding-right: 20px;
+	width: 100%;
+	text-align: center;
+}
+.container button:hover {
+	background: #fff;
+	color: #9909ac;
+}
+
+.container button:focus {
+	outline: none;
+	box-shadow: 0 0 0 4px #640670;
+	/* border: 5px solid #181d23; */
+}
